my attempts at implementing the mcv system, not done
