+++ Assets/Scenes/Battle.unity
@@ -151,6 +151,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: dc9e6cc446a09724d89609d78267266a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  unit: {fileID: 7643567009469890289, guid: 6c271d9a3a706db4cbd6c8a15a333b60, type: 3}
   armyCreator: {fileID: 290596616}
   army: []
 --- !u!114 &290596616
@@ -247,11 +248,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 1015364938}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 624a976e237bf83409842c56dc247811, type: 3}
+  m_Script: {fileID: 11500000, guid: 3e7561a068a1de145acaf0c3d1e20408, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   unit: {fileID: 7643567009469890289, guid: 6c271d9a3a706db4cbd6c8a15a333b60, type: 3}
-  ebcontroller: {fileID: 290596615}
 --- !u!114 &1015364942
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -265,72 +265,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   battleSceneCreator: {fileID: 1015364939}
---- !u!1 &1364514096
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1364514099}
-  - component: {fileID: 1364514098}
-  - component: {fileID: 1364514097}
-  m_Layer: 0
-  m_Name: EventSystem
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!114 &1364514097
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1364514096}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4f231c4fb786f3946a6b90b886c48677, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_HorizontalAxis: Horizontal
-  m_VerticalAxis: Vertical
-  m_SubmitButton: Submit
-  m_CancelButton: Cancel
-  m_InputActionsPerSecond: 10
-  m_RepeatDelay: 0.5
-  m_ForceModuleActive: 0
---- !u!114 &1364514098
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1364514096}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 76c392e42b5098c458856cdf6ecaaaa1, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_FirstSelected: {fileID: 0}
-  m_sendNavigationEvents: 1
-  m_DragThreshold: 10
---- !u!4 &1364514099
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1364514096}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 4
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1823548214
 GameObject:
   m_ObjectHideFlags: 0
