+++ Assets/Scripts/Battle/UnitManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 624a976e237bf83409842c56dc247811
+guid: 3e7561a068a1de145acaf0c3d1e20408
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
